fix: context details
