+++ hooks/chat/message-editor.ts
@@ -12,16 +12,7 @@ import { buildMessagesForAPI } from "./conversation-manager";
 import { createNewVersion, buildUpdatedVersionsList, fetchMessageVersions, updateMessageWithVersions } from "./version-manager";
 import type { MemoryStatus } from "@/types/chat";
 import { ToolStatus, MessageRole, type MessageMetadata } from "@/types/core";
-
-interface EditMessageContext {
-  messages: Message[];
-  conversationId: string | null;
-  abortSignal: AbortSignal;
-  queryClient: QueryClient;
-  onMessagesUpdate: (updater: (prev: Message[]) => Message[]) => void;
-  saveToCacheMutate: (data: { query: string; response: string }) => void;
-  onMemoryStatusUpdate?: (status: MemoryStatus) => void;
-}
+import type { EditMessageContext } from "./types";
 
 export async function handleEditMessage(
   messageId: string,
